+++ src/popup/pages/StakingTransfer/index.js
@@ -51,6 +51,7 @@ const StakingTransfer = () => {
 
   const [memo, setMemo] = useState('')
   const [feeAmount, setFeeAmount] = useState(0.1)
+  const [inputedFee, setInputedFee] = useState("")
   const [inputNonce, setInputNonce] = useState("")
   const [feeErrorTip, setFeeErrorTip] = useState("")
   const [netFeeList, setNetFeeList] = useState([])
@@ -87,7 +88,7 @@ const StakingTransfer = () => {
 
   const onFeeInput = useCallback((e) => {
     setFeeAmount(e.target.value)
-
+    setInputedFee(e.target.value)
     if (BigNumber(e.target.value).gt(10)) {
       setFeeErrorTip(i18n.t('feeTooHigh'))
     } else {
@@ -293,7 +294,8 @@ const StakingTransfer = () => {
         <AdvanceMode
           onClickAdvance={onClickAdvance}
           isOpenAdvance={isOpenAdvance}
-          feeValue={feeAmount}
+          feeValue={inputedFee}
+          feePlaceholder={feeAmount}
           onFeeInput={onFeeInput}
           feeErrorTip={feeErrorTip}